Break up format tests
