+++ README.md
@@ -31,7 +31,7 @@ The interpreter can run "hello world" in a REPL and evaluate simple expressions.
   - [ ] Support in compiler
   - [ ] Tests in compiler
 - [ ] Consider using `end` keyword for loops/conditionals
-    - I think a bare `end` would close the program/global scope?
+  - I think a bare `end` would close the program/global scope?
 - [ ] Scope Related ADRs
   - [ ] Local variables
   - [ ] Closures
